backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
argon2-cffi
pyjwt
python-multipart
google-auth
google-generativeai
requests
